separate client and server unit tests
