+++ config/jest.client.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  projects: [
+    "<rootDir>/jest/client.config.js",
+  ],
+};
