+++ config/jest.server.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  projects: [
+    "<rootDir>/jest/server.config.js",
+  ],
+};
