+++ package.json
@@ -57,7 +57,10 @@
     "start:development:withJobs": "NODE_ENV=development TS_NODE_PROJECT=./src/tsconfig.json ts-node-dev --inspect --transpile-only --no-notify -r tsconfig-paths/register --ignore-watch ./docs/ ./src/index.ts",
     "start:webpackDevServer": "ts-node --transpile-only ./scripts/start.ts",
     "start": "NODE_ENV=production node dist/index.js",
-    "test": "node --trace-warnings scripts/test.js --env=jsdom",
+    "test": "node --trace-warnings scripts/test.js --watch",
+    "test:ci": "npm run test:server && npm run test:client",
+    "test:server": "node --trace-warnings scripts/testServer.js",
+    "test:client": "node --trace-warnings scripts/testClient.js",
     "tscheck:client": "tsc --project ./src/core/client/tsconfig.json --noEmit",
     "tscheck:scripts": "tsc --project ./tsconfig.json --noEmit",
     "tscheck:server": "tsc --project ./src/tsconfig.json --noEmit",
